restore modified files in new location
