--- src/foundry/ERC20MintableByAnyone.sol
@@ -1,19 +1,0 @@
-// SPDX-License-Identifier: MIT
-
-pragma solidity 0.8.20;
-
-import "@openzeppelin/contracts/token/ERC20/ERC20.sol";
-
-/**
-   @dev This contract is used for testing only. It is a ERC20 token that can be minted by anyone.
- */
-contract ERC20MintableByAnyone is ERC20 {
-    constructor(
-        string memory name_,
-        string memory symbol_
-    ) ERC20(name_, symbol_) {}
-
-    function mint(address to, uint256 amount) public {
-        _mint(to, amount);
-    }
-}
